+++ tests/test_au/forms.py
@@ -7,5 +7,6 @@ class AustralianPlaceForm(ModelForm):
     """ Form for storing an Australian place. """
     class Meta:
         model = AustralianPlace
-        fields = ('state', 'state_required', 'state_default', 'postcode', 'postcode_required', 'postcode_default',
-                  'phone', 'name')
+        fields = ('state', 'state_required', 'state_default', 'postcode',
+                  'postcode_required', 'postcode_default',
+                  'phone', 'name', 'abn')