+++ tests/test_au/models.py
@@ -1,6 +1,5 @@
 from django.db import models
-
-from localflavor.au.models import AUPhoneNumberField, AUPostCodeField, AUStateField
+from localflavor.au.models import AUBusinessNumberField, AUPhoneNumberField, AUPostCodeField, AUStateField
 
 
 class AustralianPlace(models.Model):
@@ -12,3 +11,4 @@ class AustralianPlace(models.Model):
     postcode_default = AUPostCodeField(default="2500", blank=True)
     phone = AUPhoneNumberField(blank=True)
     name = models.CharField(max_length=20)
+    abn = AUBusinessNumberField()