+++ tests/test_au/tests.py
@@ -2,11 +2,14 @@ from __future__ import unicode_literals
 
 import re
 
+from django.core.exceptions import ValidationError
 from django.test import TestCase
 
-from localflavor.au.forms import AUPhoneNumberField, AUPostCodeField, AUStateSelect
+from localflavor.au import forms, models
+from localflavor.au.validators import AUBusinessNumberFieldValidator
 
 from .forms import AustralianPlaceForm
+from .models import AustralianPlace
 
 SELECTED_OPTION_PATTERN = r'<option value="%s" selected="selected">'
 BLANK_OPTION_PATTERN = r'<option value="">'
@@ -22,6 +25,7 @@ class AULocalflavorTests(TestCase):
              'name': 'dummy',
              'postcode': '1234',
              'postcode_required': '4321',
+             'abn': '74457506140',
              })
 
     def test_get_display_methods(self):
@@ -42,10 +46,16 @@ class AULocalflavorTests(TestCase):
         """ Test that required AUStateFields throw appropriate errors. """
         form = AustralianPlaceForm({'state': 'NSW', 'name': 'Wollongong'})
         self.assertFalse(form.is_valid())
+        self.assertEqual(set(form.errors.keys()),
+                         set(('state_required',
+                              'postcode_required',
+                              'abn')))
         self.assertEqual(
             form.errors['state_required'], ['This field is required.'])
         self.assertEqual(
             form.errors['postcode_required'], ['This field is required.'])
+        self.assertEqual(
+            form.errors['abn'], ['This field is required.'])
 
     def test_field_blank_option(self):
         """ Test that the empty option is there. """
@@ -64,7 +74,7 @@ class AULocalflavorTests(TestCase):
                                   str(self.form['postcode_required'])))
 
     def test_AUStateSelect(self):
-        f = AUStateSelect()
+        f = forms.AUStateSelect()
         out = '''<select name="state">
 <option value="ACT">Australian Capital Territory</option>
 <option value="NSW" selected="selected">New South Wales</option>
@@ -87,7 +97,7 @@ class AULocalflavorTests(TestCase):
             'abcd': error_format,
             '20001': ['Ensure this value has at most 4 characters (it has 5).'] + error_format,
         }
-        self.assertFieldOutput(AUPostCodeField, valid, invalid)
+        self.assertFieldOutput(forms.AUPostCodeField, valid, invalid)
 
     def test_AUPhoneNumberField(self):
         error_format = ['Phone numbers must contain 10 digits.']
@@ -104,4 +114,110 @@ class AULocalflavorTests(TestCase):
             '123': error_format,
             '1800DJANGO': error_format,
         }
-        self.assertFieldOutput(AUPhoneNumberField, valid, invalid)
+        self.assertFieldOutput(forms.AUPhoneNumberField, valid, invalid)
+
+    def test_abn(self):
+        error_format = ['Enter a valid ABN.']
+        valid = {
+            '53004085616': '53004085616',
+        }
+        invalid = {
+            '53 004 085 616': error_format,
+            '53004085617': error_format,
+            '5300A085616': error_format,
+        }
+        self.assertFieldOutput(forms.AUBusinessNumberField, valid, invalid)
+
+
+class AULocalFlavorAUBusinessNumberFieldValidatorTests(TestCase):
+
+    def test_no_error_for_a_valid_abn(self):
+        """Test a valid ABN does not cause an error."""
+        valid_abn = '53004085616'
+
+        validator = AUBusinessNumberFieldValidator()
+
+        validator(valid_abn)
+
+    def test_raises_error_for_abn_containing_a_letter(self):
+        """Test an ABN containing a letter is invalid."""
+        invalid_abn = '5300408561A'
+
+        validator = AUBusinessNumberFieldValidator()
+
+        self.assertRaises(ValidationError, lambda: validator(invalid_abn))
+
+    def test_raises_error_for_too_short_abn(self):
+        """Test an ABN with fewer than eleven digits is invalid."""
+        invalid_abn = '5300408561'
+
+        validator = AUBusinessNumberFieldValidator()
+
+        self.assertRaises(ValidationError, lambda: validator(invalid_abn))
+
+    def test_raises_error_for_too_long_abn(self):
+        """Test an ABN with more than eleven digits is invalid."""
+        invalid_abn = '530040856160'
+        validator = AUBusinessNumberFieldValidator()
+        self.assertRaises(ValidationError, lambda: validator(invalid_abn))
+
+    def test_raises_error_for_whitespace(self):
+        """Test an ABN can be valid when it contains whitespace."""
+        # NB: Form field should strip the whitespace before regex valdation is run.
+        invalid_abn = '5300 4085 616'
+
+        validator = AUBusinessNumberFieldValidator()
+
+        self.assertRaises(ValidationError, lambda: validator(invalid_abn))
+
+    def test_raises_error_for_invalid_abn(self):
+        """Test that an ABN must pass the ATO's validation algorithm."""
+        invalid_abn = '53004085617'
+
+        validator = AUBusinessNumberFieldValidator()
+
+        self.assertRaises(ValidationError, lambda: validator(invalid_abn))
+
+
+class AULocalFlavorAUBusinessNumberModelTests(TestCase):
+
+    def test_AUBusinessNumberModel_invalid_abn_raises_error(self):
+
+        place = AustralianPlace(**{
+            'state': 'WA',
+            'state_required': 'QLD',
+            'name': 'dummy',
+            'postcode': '1234',
+            'postcode_required': '4321',
+            'abn': '5300 4085 616 INVALID',
+        })
+
+        self.assertRaises(ValidationError, place.clean_fields)
+
+
+class AULocalFlavourAUBusinessNumberFormFieldTests(TestCase):
+
+    def test_abn_with_spaces_remains_unchanged(self):
+        """Test that an ABN with the formatting we expect is unchanged."""
+
+        field = forms.AUBusinessNumberField()
+
+        self.assertEqual('53 004 085 616', field.prepare_value('53 004 085 616'))
+
+    def test_spaces_are_reconfigured(self):
+        """Test that an ABN with formatting we don't expect is transformed."""
+
+        field = forms.AUBusinessNumberField()
+
+        self.assertEqual('53 004 085 616', field.prepare_value('53004085616'))
+        self.assertEqual('53 004 085 616', field.prepare_value('53 0 04 08561 6'))
+
+
+class AULocalFlavourAUBusinessNumberModelFieldTests(TestCase):
+
+    def test_to_python_strips_whitespace(self):
+        """Test the value is stored without whitespace."""
+
+        field = models.AUBusinessNumberField()
+
+        self.assertEqual('53004085616', field.to_python('53 004 085 616'))